Mac ÜC: Enable GPU back-pressure from the browser to the renderer
When the browser compositor does a SwapBuffers to its output surface,
take a lock on the compositor until that swap is displayed on the screen.
This prevents the renderer from running ahead of the browser.

If the renderer sends a new frame (it can send at most one), then track
this event, and request redraw as soon as the lock is released.

BUG=314190

Review URL: https://codereview.chromium.org/316103004