Clean up shill fake implementations
This is in preparation of merging NetworkState and FavoriteState,
crbug.com/375955.

The merge will involve a number of changes to ShillPropertyHandler
and in the process of making those changes some additional logging
and test cleanup was necessary that isn't directly related to the
merge itself.

BUG=none
For trivial changes to chrome/browser/ui/webui/options/
TBR=dbeam@chromium.org

Review URL: https://codereview.chromium.org/299403012