[Cocoa] Clean-up: Update comments regarding drawRect override in TabContentsContainerView.
Once upon a time, in an era of great suffering, Chromium Mac was not
using CoreAnimation, with it being necessary to override -drawRect: to
paint the tab contents area background fill.  That time has, unfortunately,
not yet passed.

This change simply updates a few comments to wrap up the resolution
of bug 354598, which ccameron@ has mainly fixed in an earlier
change; and also defaults the bgColor to white when the themed color
is not available.

BUG=354598

Review URL: https://codereview.chromium.org/319703006