WebRTC: Update isolate file location for audio_decoder_unittests
The new file is already in place and the old one is just a
copy of it left there waiting for this CL to land before it
can be removed.

I'm setting NOTRY since this is not used by any Chromium 
bot (only WebRTC-maintained bots).

NOTRY=True
TBR=frankf@chromium.org
BUG=webrtc:3462

Review URL: https://codereview.chromium.org/323833002