Removing WTL from content runner.
The _Module object is used by ATL COM classes to lock the module, like so

_pAtlModule->Lock();

For example when CreateInstance is called:

BrowserAccessibility* BrowserAccessibility::Create() {
  CComObject<BrowserAccessibilityWin>* instance;
  HRESULT hr = CComObject<BrowserAccessibilityWin>::CreateInstance(&instance);

We need to get rid of WTL so we can get rid of ATL since it
causes many problems. See bug for details.

BUG=5027
TEST=chrome runs. Accessibility works.

Review URL: https://codereview.chromium.org/323723002