rendering_helper - Replace multiple windows by a full-screen window.
On some ChromeOS devices, there would be extra effert to copy content
from a window to the screen when the window is not full-sized.

This change tries to replace the windows by a full-screen window and
let different decoders draw on different areas of the window.

This did improve the performance a lot. After this change, the time
took by the thumbnail test is improved from 7 seconds to 1.6 seconds.

Due to crbug/270064, we cannot run two fullscreen X apps. The chrome
should be stopped before unning the vda_unittest.

    # touch /var/run/disable_chrome_restrat
    # restart ui

BUG=chrome-os-partner:28176
TEST=Run the test on peach-pit

Review URL: https://codereview.chromium.org/271593010