Mojo: cleanup redirect handling for network service.
Eliminate redirects field. Leave it up to clients to log the redirect chain if
so desired. It is not worth sending this list over the pipe each time we see
a redirect.

Add support for "follow_redirects" request parameter. Rename it with an "auto_"
suffix for greater clarity.

R=davemoore@chromium.org

Review URL: https://codereview.chromium.org/324023002