Files.app: Refresh metadata one by one when the directory contents are changed
Previously, when the directory contents are changed, once all the metadata were invalidated and then new metadata were fetched. But it caused flicker between invalidation and re-fetching.

This patch introduces the new method to force retrieving the metadata. And with this method, the filelist refreshes the metadata one by one and flicker is not occurred.

BUG=373629
TEST=manually
R=hirono@chromium.org

Review URL: https://codereview.chromium.org/317833003