cc: Don't cleanup the pending twin's low res tiling.
After drawing, we cull unused tilings on the active tree. If we cull
the tiling which is the low res tiling on the pending tree, and then
later do a pinch such that the active tree wants a tiling of that scale
again, it would create the tiling which is already on the pending tree
and cause a bad state.

Adds a unit test that was hitting the CHECK() in PictureLayerImpl when
it changed the scale of the active layer to 1.f to match the pending
layer after drawing. Now it verifies that the active layer keeps the
pending low res tiling.

R=enne
BUG=358350

Review URL: https://codereview.chromium.org/323193002