Fix chrome.extension.getViews returns multiple popup windows.
When we invoke +[ExtensionPopupController showURL:...],
we close the old Popup using [gPopup close],but we don't
release it here. So host_ is only released 
when -[ExtensionPopupController windowWillClose] is invoked.
At this time, the new popup might be ready. And when we use
chrome.extension.getViews or some other apis we might get
the dying window.
So move shutdown of host_ from -[ExtensionPopupController windowWillClose]
into -[ExtensionPopupController close].

BUG=376511

Review URL: https://codereview.chromium.org/321093002