Prevent incorrect start time from being used during new range creation.
If a SourceBuffer::Remove() call removed the GOP that is currently being
appended, an incorrect start timestamp would be used when the next GOP was
appended. This ultimately led to a crash if a followup Remove() call
ended up deleting the range. This change fixes the code that determines
the start timestamp for the range created for the second GOP.

BUG=382815
TEST=SourceBufferStreamTest, Remove_WholeGOPBeingAppended

Review URL: https://codereview.chromium.org/324223003