Disable Warning 4702 for the PGO builds.
There's more optimization done on a PGO build than on a regular one, this results in some "Unreachable code" warnings that we don't see normally.

But probably anything that this would catch that wouldn't be caught in a normal build isn't going to actually be a bug, so the incremental value of C4702 for PGO builds is likely very small and blocks the setup of the PGO bots.

BUG=380175

Review URL: https://codereview.chromium.org/315323002