Makes the view manager notify the client of new roots correctly
Prior to this patch the server would not notify the client of a new
root if the client already knew about the root. Now it always
notifies.

Here's an example of this case: two nodes, A and B with B a child of
A and A the root for the client. If B is added as a root then
previously I would not send B in OnRootsAdded. Now I do.

BUG=365012
TEST=none
R=ben@chromium.org

Review URL: https://codereview.chromium.org/323413003