Do not skip subtrees with wheel or touch handlers
When calculating draw properties, we attempt to skip subtrees whenever
we can, but we missed a corner case: even though we don't necessarily
draw a layer with a wheel or touch handler, it will participate in hit
testing and must therefore have up-to-date transforms, targets, etc.
In the same way that we've added logic to prevent skipping subtrees
containing layers with copy requests, this CL prevents skipping subtrees
with handlers.

R=danakj@chromium.org
BUG=377738

Review URL: https://codereview.chromium.org/335633003