cc: Move tiling management out of draw properties calculation.
This is mainly plumbing code.
As part of the changes, we are moving ManageTiling as part of 
UpdateTilePriorities.
We keep picturelayers content scale as 1,and maintain the remaining 
required scales in draw properties. This scales we apply later, 
to shared quad state and draw transform during AppendQuads.


BUG=None

Review URL: https://codereview.chromium.org/271533011