Cast: Synthetic benchmark tool.
This benchmark tests how much latency, packet loss and bandwidth loss we can tolerate.
It runs hundreds of times faster than real-time and results are independent of CPU speed.
So far, it does a series of binary searches to find a spanning tree of the space in which
our algorithms can operate comfortably.

This is a re-upload of cl 293173002, which was reverted because it broke some windows build.

Committed: https://src.chromium.org/viewvc/chrome?view=rev&revision=276603

Review URL: https://codereview.chromium.org/308713005