mac: Add overscroll animator slider.
An overscroll animator performs animations related to overscrolling for the
WebContentsView. I added a protocol for overscroll animators on mac, and wrote
a single implementation OverscrollAnimatorSliderView. This implementation
performs an overscroll animation by sliding the current RenderWidgetHostView
off the screen to the right (for backwards navigation), or by sliding an image
of the next page onto the screen from the right (for forwards navigation).
Neither the implementation nor the protocol are hooked into WebContentsView
yet.

BUG=381466

Review URL: https://codereview.chromium.org/314393003