Added a pref to hide the Managed Bookmarks.
If the user has managed bookmarks configured via policy then there is a new
option in the bookmark bar context menu that can be used to hide the
managed bookmarks folders, similar to the option to hide the apps shortcut.
However this option is not present if there is no managed bookmarks policy
present.

BUG=49598

Review URL: https://codereview.chromium.org/328793004