Ensure HandleDocumentLoad isn't called if the instance is destroyed
This adds a check to PepperPluginInstanceImpl::DidDataFromWebURLResponse
to avoid calling the instances HandleDocumentLoad function if the plugin
instance is deleted before the function gets called.

It also adds a check to PepperURLLoaderHost::Close to avoid trying to access
an instance that has already been destroyed. A TODO is also added because
the behavior of PepperURLLoaderHost::Close is incorrect.

BUG=384197,372548

Review URL: https://codereview.chromium.org/335463006