Support casting for embedded YT videos
This is part 1 of the code to start supporting embedded YT for cast. The main
issue is that YT embedded videos need to be identified via the frame that
contains it. This CL sets the frame URL of each video in the MediaPlayerAndroid
object.

NOTRY=true

Clank downstream piece to come later. 
Min / Aaron: will send you the doc I have that explains this in more detail.

BUG=358573

Review URL: https://codereview.chromium.org/302453012