Add testRunner.capturePixelsThen() method
This method is similar to displayAsyncThen(), but allows the caller
to get the captured pixel data.

This is to be used in repaint tests to produce the minimum repaint
baselines. This is useful to discover under-repaint bugs with the help
of repaint visualization in layout test results, and to estimate
over-repaints.

R=enne
TBR=dpranke

Review URL: https://codereview.chromium.org/301243022