Prevent default_apps external downloads from being auto-removed.
The default_apps::Provider is an extensions::ExternalProviderImpl
which provides extensions for installation on first-run of Chrome.
After the first run, it behaves as if it has no extensions to provide.

Due to this, any extensions it provides must not be marked as coming
from an external location, otherwise they will be removed on the
next run of Chrome (since Chrome will verify that each external
extension is known by at least one external provider and this will
fail.)

To fix this problem, default_apps::Provider now marks its
externally downloaded extensions (which come only from the Web Store)
as having an INTERNAL location.

BUG=384300,384060
R=asargent@chromium.org

Review URL: https://codereview.chromium.org/335233002