Restore non surface gpu memory allocation behavior
When using surfaces the browser compositor's context is considerd
"offscreen" and thus uses the memory limited computed for contexts without
a surface. This changed from 64MB to 1 byte in
https://src.chromium.org/viewvc/chrome?revision=277056&view=revision since
the other users of this number were using it as a boolean value. This restores
the old behavior and simplifes the computation a bit for non-Android platforms.

R=ccameron@chromium.org

Review URL: https://codereview.chromium.org/333313006