VAVDA: Don't use an event to wait for the decoder thread to be drained.
On destruction, the ChildThread waits for a Signal() to a WaitableEvent
posted onto the decoder thread, to ensure that the thread is drained.
This is not needed however, as Stop() will ensure that the thread is
drained anyway.

But because the Signal() task is posted onto the thread via a proxy, if
the thread is already stopped, it will let us post, but the task will
never execute, and we could wait forever.

BUG=384660
TEST=vdatest,video playback

Review URL: https://codereview.chromium.org/337853002