Remove unneeded shm versions of bucket functions from command buffer
Remove shm versions of bucket functions from command buffer for commands
that do not need two different ways to send pointer data.

The actual code uses only bucket commands. The shm versions are
either unused or used only by the unit tests.

Make the unit tests use the immediate commands.

Solves the problem of not having to implement one useless function for
the a new command when using bucket commands.

Support the case of CompressedTex{,Sub}Image2D, where the function
is implemented with shm and bucket commands.

In the code generator function info table, replace 'immediate': False
and 'bucket': True properties of function info objects with
'data_transfer_method': ['immediate', 'bucket', 'shm' ] property.

Remove the handwritten GetAttribLocation command and rename
GetAttribLocationBucket to GetAttribLocation. This is consistent with
other handwritten commands: the command name is the gl call name, and it
just happens to use buckets to communicate some of its arguments.  The
command does not have 'Bucket' prefix due to it not being a 'Bucket'
variant. Do the same with GetUniformLocation and
GetUniformLocationBucket.

BUG=373763

Review URL: https://codereview.chromium.org/282253002