Identity API: add chrome.identity.getProfileUserInfo
getProfileUserInfo returns the email address and obfuscated gaia id of
the user signed into a profile. If the user is not signed in, empty
strings are returned.

This API is different from identity.getAccounts in two ways. The information returned is available offline, and it only applies to the primary account for the profile (even when multiple login is enabled).

BUG=368343

Review URL: https://codereview.chromium.org/327403003