Do not enable network technology if not available
We currently set a technology to 'enabling' even if it is not available.
We also enable / disable both Cellular and Wimax on the same UI element
since both are never expected to be available.
Simple fix is to ignore technologies that are not available.

BUG=368754

Review URL: https://codereview.chromium.org/336313002