[NaCl SDK] Fix create_nmf on dynamic bionic nexes
create_nmf runs objdump on its shared objects to determine what
dependencies it has. libc.so is a dependency of bionic, and it is an ELF
file. glibc has libc.so as well, but it is not an ELF file, but a linker
script.

There was a check in get_shared_deps.py that prevented running objdump on
libc.so, if "bionic" is not in the path. Since we've added a package
called "bionic_canary", "bionic" is always in the path, so objdump was
being run on glibc's "libc.so", which fails.

The (hack) fix is to check if a relative path from the root of the SDK has
"bionic". This should only be true if the path is
"toolchain/linux_arm_bionic/...".

BUG=none
R=sbc@chromium.org, noelallen@chromium.org
TEST=test_all.py

Review URL: https://codereview.chromium.org/341983002