Send revision and webkit-revision to runtest.py during sizes
Currently, the sizes test on master.chromium bots inspects the filesystem to get
the revision that it includes in the json it uploads (see runtest.py
_GetSvnRevision).  kbr@ landed a change that allows runtest.py to take
--revision and --webkit-revision as command line arguments, in which case it
will trust those and not inspect the filesystem.

This change causes the sizes step to pass the got_*_revision properties into
the runtest.py invocation, so that it won't get it wrong when we are pretending
to be SVN but actually sitting on top of a git checkout.

R=eglaysher@chromium.org, hinoka@chromium.org, stip@chromium.org

Review URL: https://codereview.chromium.org/343813003