Adds USB writing for OS X.
To support imageWriterPrivate on Mac for the Chromebook Recovery Tool we will have to write to USB drives on Mac.  OS X allows us to open a single FD with elevated privileges through running the authopen command.

Unmounting and claiming the disk are a bit more complicated as they require use of the Disk Arbitration framework.

Disk Arbitration is an asynchronous interface.  In order to simplify the code we use a mutex to block.

The core logic is derived from the previous recovery tool.  See googleclient/chrome/chromeos_recovery/mac/ChromeOSRecovery/AppController.mm in the internal Google repositories.

BUG=376398
BUG=318451

Review URL: https://codereview.chromium.org/294163008