Ignore EnterNotify and LeaveNotify events from NativeViewGlSurfaceGLX's window. https://codereview.chromium.org/296003010 added a child window to the toplevel X window. As a result of the CL, an EnterNotify event is sent immediately after each mouse press. WindowTreeHostX11:Dispatch() transforms the EnterNotify to a synthetic mouse move.
In terms of HTML events, the CL causes an 'onmousemove' event to be fired
immediately after the 'onmousedown' event. This differs from the previous
behavior where the initial 'onmousemove' event was always fired some
indeterminate time after the 'onmousedown'.

For the sake of safety, this CL suppresses the extra EnterNotify.

BUG=385716
TEST=None

Review URL: https://codereview.chromium.org/346673002