device/bluetooth: Send GattServiceChanged event for new Chrc. properties.
BluetoothRemoteGattServiceChromeOS now sends a GattServiceChanged observer event
when a new value for a characteristic property is received. This is needed when
the "Characteristic Extended Properties" descriptor is present for a
characteristic and its value is asynchronously read by bluetoothd, which results
in a PropertiesChanged signal for the "Flags" property of GattCharacteristic1.

BUG=382253

Review URL: https://codereview.chromium.org/346743002