Reland change to use a native MessagePump instead of a MessagePumpDefault. Original CL: https://codereview.chromium.org/331513002/
The previous CL caused a bunch of LibJingle unit tests to fail because the tests
were using a gtest death test which runs in the context of a fork()ed but not
execu()ed process.

A CF Message loop can't live in this environment.  Explicitly construct a
MessagePumpDefault for these tests.

BUG=356804,385604

Review URL: https://codereview.chromium.org/331983002